refactor: 일부 클래스, 변수명 명확히 변경, customizer 분리
