+++ common/src/main/java/com/pickcar/presentation/advice/ErrorResponseAdvice.java
@@ -1,11 +1,11 @@
 package com.pickcar.presentation.advice;
 
-import com.pickcar.presentation.dto.response.ErrorResponse;
-import com.pickcar.exception.GlobalException;
-import org.springframework.http.ResponseEntity;
 import com.pickcar.constants.GlobalStatic;
+import com.pickcar.exception.GlobalException;
+import com.pickcar.presentation.dto.response.ErrorResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
